Complted deal related code works and postman testing

+++ src/Application/Deals/Commands/DeleteDeal/DeleteDealCommand.cs
@@ -26,6 +26,7 @@ public class DeleteDealCommandHandler : IRequestHandler<DeleteDealCommand, int>
     public async Task<int> Handle(DeleteDealCommand request, CancellationToken cancellationToken)
     {
         var deal = await _context.Deals.FindAsync(request.Id);
+        if (deal == null) return 0;
         _context.Deals.Remove(deal);
         return await _context.SaveChangesAsync(cancellationToken);
